requirements.txt
aiogram==3.1.1
aiosqlite==0.19.0
python-dotenv==1.0.0
qrcode==7.4.2
requests==2.31.0
asyncio==3.4.3
pillow==10.0.0
psutil==5.9.5
